refactor: Move some filesystem helpers into the fs module
